==> 1-statistical_model.ipynb <==
# Statistical Modelling of total contacts: model fit
#### Non negative binomial model with EPID random effect

## Models for Adult respondents

# # Adults
library(colorspace)
library(MASS)
library(glmmTMB)
library(dplyr)
library(vctrs)
library(ggplot2)
library(shape)
library(stringr)
library(RColorBrewer)
library(huxtable)
library(texreg)
library(tidyverse)
library(ggstats)
library(ggforce)
library(ggstance)
library(broom.mixed)
# library(Polychrome)
library(Hmisc)
library(showtext)
library(haven)

font_add_google(name = "Barlow", family = "Barlow")
showtext_auto()

# # ```{r data}
load("data_in/datamod.Rdata")

# ```{r adults}
data_for_model = cbind(data_mod[,c("total_contacts", "total_contacts_nohh", "total_contacts_prol", "total_contacts_prol_soft", "total_contacts_nohh_prol","wave","EPID")],data_mod[,1:34])
names(data_for_model)[1:7]=c("total_contacts", "total_contacts_nohh", "total_contacts_prol", "total_contacts_prol_soft","total_contacts_nohh_prol","wave","EPID")

### Prepare "data_fo_model_prol"
data_for_model_prol <- data_for_model[data_for_model$total_contacts_prol <= 100,] # censoring at 100
Hmisc::label(data_for_model_prol) =  as.list(var.labels[match(names(data_for_model_prol), names(var.labels))])

colnames(data_for_model)

# A1) direct+indirect contacts with cohabitants
adults_indirect_totcont = glmmTMB(total_contacts_prol ~ wave + (1|EPID) +
                                age_group_10+
                                education+
                                occupation_agg+
                                income_threecat+
                                attend_in_presence+
                                sunday+
                                d_vacc2+
                                d_time_since_covid + 
                                chronic_comorb_self2 +
                                chronic_comorb_cohab2+
                                d_senior65+
                                d_children +
                                hh_size,
                              family=nbinom2,
                              control=glmmTMBControl(optimizer=optim,optArgs=list(method="BFGS")),
                         data=data_for_model_prol) 
summary(adults_indirect_totcont)
# Save model fit to RDS
saveRDS(adults_indirect_totcont, file = "model_out/adults_indirect_totcont_glmmTMB.rds")

# A2) direct+indirect contacts without cohabitants
adults_indirect_ncohabcont = glmmTMB(total_contacts_nohh_prol ~ wave + (1|EPID) +
                                respondent_gender+
                                age_group_10+
                                education+
                                occupation_agg+
                                income_threecat+
                                attend_in_presence+
                                sunday+
                                d_vacc2+
                                d_time_since_covid + 
                                chronic_comorb_self2 +
                                chronic_comorb_cohab2+
                                d_senior65+
                                d_children +
                                hh_size,
                       family=nbinom2,
                       control=glmmTMBControl(optimizer=optim,optArgs=list(method="BFGS")),
                       data=data_for_model_prol)
summary(adults_indirect_ncohabcont)
# Save model fit to RDS
saveRDS(adults_indirect_ncohabcont, file = "model_out/adults_indirect_ncohabcont_glmmTMB.rds")

# B1) direct contacts (only) with cohabitants
adults_orig_totcont = glmmTMB(total_contacts ~ wave + (1|EPID) +
                                respondent_gender+
                                age_group_10+
                                education+
                                occupation_agg+
                                income_threecat+
                                attend_in_presence+
                                sunday+
                                d_vacc2+
                                d_time_since_covid + 
                                chronic_comorb_self2 +
                                chronic_comorb_cohab2+
                                d_senior65+
                                d_children +
                                hh_size,
                        control=glmmTMBControl(optimizer=optim,optArgs=list(method="BFGS")),
                        family=nbinom2,
                        data=data_for_model_prol)
summary(adults_orig_totcont)
# Save model fit to RDS
saveRDS(adults_orig_totcont, file = "model_out/adults_orig_totcont_glmmTMB.rds")

# B2) direct+indirect soft contacts with cohabitants
adults_soft_totcont = glmmTMB(total_contacts_prol_soft ~ wave + (1|EPID) +
                                respondent_gender+
                                age_group_10+
                                education+
                                occupation_agg+
                                income_threecat+
                                attend_in_presence+
                                sunday+
                                d_vacc2+
                                d_time_since_covid + 
                                chronic_comorb_self2 +
                                chronic_comorb_cohab2+
                                d_senior65+
                                d_children +
                                hh_size,
                        control=glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")),
                        family=nbinom2,
                        data=data_for_model_prol)
summary(adults_soft_totcont)
# Save model fit to RDS
saveRDS(adults_soft_totcont, file = "model_out/adults_soft_totcont_glmmTMB.rds")

## Models for Underage respondents

# # Children
# ```{r children}
data_mod = data_mod_children %>% 
  rename(parental_education = education,
         parental_occupation= occupation_agg2
         )
data_mod$parental_occupation[data_mod$respondent_age>15]=NA #set to NA for 16-17 they answered about themselves
data_mod$parental_occupation=ifelse(data_mod$parental_occupation %in% c("Employed","Student"),"Employed","Other (not working)")
data_mod$kindergarten_05[data_mod$respondent_age>5]="No"
data_mod$hh_size = droplevels(data_mod$hh_size)
data_mod <- arrange(data_mod,EPID,wave)
data_mod$presence_school <- ifelse(data_mod$presence_school == "In presence", "In presence", "Dind't attend")
data_mod$presence_school <- factor(data_mod$presence_school)

data_mod$age_group_child <- ifelse(data_mod$respondent_age <= 5, "0-5 y", 
                                   ifelse(data_mod$respondent_age > 5 & data_mod$respondent_age <=13, "6-13 y",
                                          ifelse(data_mod$respondent_age >13, "14-17 y", NA)))
data_mod$age_group_child <- factor(data_mod$age_group_child, levels = c("0-5 y", "6-13 y", "14-17 y"))
data_mod$age_group_child <- relevel(data_mod$age_group_child, ref = 2)

var.labels = c(var.labels,
               age_group_child = "Age group of the child",
               parental_occupation = "Parental occupation",
               parental_education = "Parental education",
               presence_school = "Went to school is presence",
               hh_size = "Household size")
Hmisc::label(data_mod) =  as.list(var.labels[match(names(data_mod), names(var.labels))])

data_for_model_prol <- data_mod[data_mod$total_contacts_prol <= 100 & data_mod$total_contacts_prol!=0,] # censoring at 100

colnames(data_mod)

# A) children direct+indirect contacts
children_indirect_totcont = glmmTMB(total_contacts_prol ~ wave + (1|EPID) +
                                  respondent_gender+
                                  age_group_child +
                                  parental_education +
                                  parental_occupation+
                                  income_threecat+
                                  presence_school +
                                  kindergarten_05 +
                                  sunday+
                                  d_time_since_covid +
                                  d_vacc2 +
                                  other_vacc +
                                  chronic_comorb_self2+
                                  chronic_comorb_cohab2+
                                  d_senior65+
                                  hh_size
                         + d_vacc2:wave,
                       family=nbinom2,
                       data=data_for_model_prol)
summary(children_indirect_totcont)
# Save model fit to RDS
saveRDS(children_indirect_totcont, file = "model_out/children_indirect_totcont_glmmTMB.rds")

# A2) children direct+indirect contacts without cohabitants
children_indirect_ncohabcont = glmmTMB(total_contacts_nohh_prol ~ wave + (1|EPID) +
                          respondent_gender+
                                  age_group_child +
                                  parental_education +
                                  parental_occupation+
                                  income_threecat+
                                  presence_school +
                                  kindergarten_05 +
                                  sunday+
                                  d_time_since_covid +
                                  d_vacc2 +
                                  other_vacc +
                                  chronic_comorb_self2+
                                  chronic_comorb_cohab2+
                                  d_senior65+
                                  hh_size
                         + d_vacc2:wave,
                       family=nbinom2,
                       data=data_for_model_prol)
summary(children_indirect_ncohabcont)
# Save model fit to RDS
saveRDS(children_indirect_ncohabcont, file = "model_out/children_indirect_ncohabcont_glmmTMB.rds")

# B1) children direct contacts (only) with cohabitants
children_orig_totcont = glmmTMB(total_contacts ~ wave + (1|EPID) +
                          respondent_gender+
                                  age_group_child +
                                  parental_education +
                                  parental_occupation+
                                  income_threecat+
                                  presence_school +
                                  kindergarten_05 +
                                  sunday+
                                  d_time_since_covid +
                                  d_vacc2 +
                                  other_vacc +
                                  chronic_comorb_self2+
                                  chronic_comorb_cohab2+
                                  d_senior65+
                                  hh_size
                         + d_vacc2:wave,
                       family=nbinom2,
                       data=data_for_model_prol)
summary(children_orig_totcont)
# Save model fit to RDS
saveRDS(children_orig_totcont, file = "model_out/children_orig_totcont_glmmTMB.rds")

# B2) children direct+indirect soft contacts (only) with cohabitants
children_soft_totcont = glmmTMB(total_contacts_prol_soft ~ wave + (1|EPID) +
                          respondent_gender+
                                  age_group_child +
                                  parental_education +
                                  parental_occupation+
                                  income_threecat+
                                  presence_school +
                                  kindergarten_05 +
                                  sunday+
                                  d_time_since_covid +
                                  d_vacc2 +
                                  other_vacc +
                                  chronic_comorb_self2+
                                  chronic_comorb_cohab2+
                                  d_senior65+
                                  hh_size
                         + d_vacc2:wave,
                       family=nbinom2,
                       data=data_for_model_prol)
summary(children_soft_totcont)
# Save model fit to RDS
saveRDS(children_soft_totcont, file = "model_out/children_soft_totcont_glmmTMB.rds")
